requirements.txt
numpy==1.16.6
opencv-python==3.4.1.15
tensorflow-gpu==2.0.0
keras==2.3.0


